componentize select all
